requirements.txt
# # # --- Core FastAPI Dependencies ---
# # fastapi
# # uvicorn[standard]       # production-grade ASGI server
# # python-dotenv           # for env file support
# # aiofiles

# # # --- Async HTTP & Requests ---
# # aiohttp>=3.8.0
# # requests>=2.28.0

# # # -Azure deployment--
# # azure-functions

# # # --- AWS and Lambda Utilities (if used) ---
# # boto3>=1.26.0
# # aws-lambda-powertools>=2.8.0
# # mangum                  # ASGI adapter for AWS Lambda (optional on Azure)

# # # --- AI Agent Framework (Autogen) ---
# # autogen-core
# # autogen-agentchat
# # autogen-ext[openai]     # ensure OpenAI support is pulled in

# # # --- Optional Tools ---
# # Pillow>=9.0.0           # used if image generation or markdown-to-pdf has images

# # pandas==2.0.3
# # numpy==1.24.3
# # yfinance==0.2.28
# # scikit-learn==1.3.0
# # xgboost==1.7.6
# # lightgbm==4.0.0
# # matplotlib==3.7.2
# # seaborn==0.12.2
# # streamlit==1.25.0
# # flask==2.3.2
# # plotly==5.15.0

# # --- Core FastAPI Dependencies ---
# fastapi
# uvicorn[standard]
# python-dotenv
# aiofiles

# # --- Async HTTP & Requests ---
# aiohttp>=3.8.0
# requests>=2.28.0

# # --- Azure deployment ---
# azure-functions

# # --- AWS and Lambda Utilities ---
# boto3>=1.26.0
# aws-lambda-powertools>=2.8.0
# mangum

# # --- AI Agent Framework (Autogen) ---
# autogen-core
# autogen-agentchat
# autogen-ext[openai]

# # --- Image Processing ---
# Pillow>=9.0.0

# # --- Data Science & ML (Updated versions) ---
# pandas>=2.0.0,<2.2.0    # Use compatible range instead of exact version
# numpy>=1.24.0,<1.27.0   # Compatible range
# yfinance>=0.2.28
# scikit-learn>=1.3.0
# xgboost>=1.7.0
# lightgbm>=4.0.0
# matplotlib>=3.7.0
# seaborn>=0.12.0
# plotly>=5.15.0

# # --- Web Apps ---
# streamlit>=1.25.0
# flask>=2.3.0

# # --- Additional for build tools ---
# setuptools>=68.0.0
# wheel>=0.40.0
# pyautogen

streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
python-dotenv>=1.0.0
yfinance>=0.2.18

# AutoGen dependencies (optional)
autogen-core>=0.2.0
autogen-agentchat>=0.2.0
autogen-ext[openai]>=0.2.0

# Google Generative AI (optional)
google-generativeai>=0.3.0

xgboost
scikit-learn

langchain
langchain-community
langchain-openai
chromadb
sentence-transformers
pypdf
unstructured
tiktoken
faiss-cpu
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
scipy>=1.11.0


# Nexus Sphere Advanced RAG System Requirements
# Install with: pip install -r requirements_rag.txt

# Core RAG Dependencies
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.20

# Vector Store
qdrant-client>=1.7.0
chromadb>=0.4.22  # Alternative vector store

# Embeddings
sentence-transformers>=2.2.2

# Document Processing
pymupdf>=1.23.8  # PyMuPDF for PDF processing
pypdf>=3.17.4  # Alternative PDF library
beautifulsoup4>=4.12.2  # HTML parsing
lxml>=4.9.3  # XML parsing
python-magic>=0.4.27  # File type detection
unstructured[pdf]>=0.11.0  # Advanced document parsing (optional)

# Web Scraping
aiohttp>=3.9.1  # Async HTTP client
aiofiles>=23.2.1  # Async file operations
requests>=2.31.0
html5lib>=1.1

# Text Processing
tiktoken>=0.5.2  # Token counting for OpenAI models
spacy>=3.7.2  # NLP (optional, for entity extraction)
# python -m spacy download en_core_web_sm  # Run after installing spacy

# Search & Ranking
rank-bm25>=0.2.2  # BM25 keyword search
faiss-cpu>=1.7.4  # Alternative to Qdrant (optional)

# Utilities
ratelimit>=2.2.1  # Rate limiting
python-dotenv>=1.0.0  # Environment variables
tqdm>=4.66.1  # Progress bars
numpy>=1.24.3
pandas>=2.0.0

# Data Validation
pydantic>=2.5.0

# Async Support
nest-asyncio>=1.5.8

# Caching
diskcache>=5.6.3

# Logging
loguru>=0.7.2  # Better logging (optional)

# Testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Development (optional)
black>=23.12.0  # Code formatter
flake8>=6.1.0  # Linter
mypy>=1.7.1  # Type checker

# Nexus Sphere Advanced RAG System Requirements
# NO DOCKER REQUIRED!

# Core RAG Dependencies
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.20

# Vector Stores (NO DOCKER NEEDED)
chromadb>=0.4.22  # Primary - persistent, no Docker
faiss-cpu>=1.7.4  # Alternative - fastest search

# Embeddings
sentence-transformers>=2.2.2

# Document Processing
pymupdf>=1.23.8  # PyMuPDF for PDF processing
pypdf>=3.17.4  # Alternative PDF library
beautifulsoup4>=4.12.2  # HTML parsing
lxml>=4.9.3  # XML parsing
unstructured[pdf]>=0.11.0  # Advanced document parsing (optional)

# Web Scraping
aiohttp>=3.9.1  # Async HTTP client
aiofiles>=23.2.1  # Async file operations
requests>=2.31.0
html5lib>=1.1

# Text Processing
tiktoken>=0.5.2  # Token counting
spacy>=3.7.2  # NLP (optional)

# Search & Ranking
rank-bm25>=0.2.2  # BM25 keyword search

# Utilities
ratelimit>=2.2.1
python-dotenv>=1.0.0
tqdm>=4.66.1
numpy>=1.24.3
pandas>=2.0.0
pydantic>=2.5.0

# Async Support
nest-asyncio>=1.5.8

# Testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
langchain-huggingface
sec-api